Remove Unneeded Usings
The usings were causing the default static HttpHandler to become disposed prematurely.

Restore ENvironment variable setting to baseprojectmanager.
