+++ src/Shared-lib/DefaultHttpClientFactory.cs
@@ -5,9 +5,23 @@ namespace Microsoft.CST.OpenSource
     using System;
     using System.Net.Http;
 
+    /// <summary>
+    /// This is the HttpClientFactory that will be used if one is not specified.  This factory lazily constructs a single HttpClient and presents it for reuse to reduce resource usage.
+    /// </summary>
     public sealed class DefaultHttpClientFactory : IHttpClientFactory
-    {            // Initialize the static HttpClient
-        private static SocketsHttpHandler handler = new()
+    {
+
+        public DefaultHttpClientFactory(string? userAgent)
+        {
+            _httpClientLazy = new(() =>
+            {
+                HttpClient cli = new(handler);
+                cli.DefaultRequestHeaders.UserAgent.ParseAdd(userAgent ?? "OSSDL");
+                return cli;
+            });
+        }
+
+        private static readonly SocketsHttpHandler handler = new()
         {
             AllowAutoRedirect = true,
             UseCookies = false,
@@ -17,9 +31,13 @@ namespace Microsoft.CST.OpenSource
             AutomaticDecompression = System.Net.DecompressionMethods.All
         };
 
-        private static readonly Lazy<HttpClient> _httpClientLazy =
-            new Lazy<HttpClient>(() => new HttpClient(handler));
+        private readonly Lazy<HttpClient> _httpClientLazy;
 
+        /// <summary>
+        /// Returns the singleton HttpClient for this factory.
+        /// </summary>
+        /// <param name="name"></param>
+        /// <returns></returns>
         public HttpClient CreateClient(string name) => _httpClientLazy.Value;
     }
 }
