+++ src/Shared-lib/PackageManagers/BaseProjectManager.cs
@@ -16,9 +16,13 @@ namespace Microsoft.CST.OpenSource.PackageManagers
     using System.Threading.Tasks;
     using Utilities;
     using Version = SemanticVersioning.Version;
+    using Microsoft.CST.OpenSource.Helpers;
 
     public class BaseProjectManager
     {
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0044:Add readonly modifier", Justification = "Modified through reflection.")]
+        private string ENV_HTTPCLIENT_USER_AGENT = "OSSDL";
+
         /// <summary>
         /// Initializes a new instance of the <see cref="BaseProjectManager"/> class.
         /// </summary>
@@ -31,9 +35,10 @@ namespace Microsoft.CST.OpenSource.PackageManagers
 
         public BaseProjectManager(string destinationDirectory)
         {
+            EnvironmentHelper.OverrideEnvironmentVariables(this);
             Options = new Dictionary<string, object>();
             TopLevelExtractionDirectory = destinationDirectory;
-            HttpClientFactory = new DefaultHttpClientFactory();
+            HttpClientFactory = new DefaultHttpClientFactory(ENV_HTTPCLIENT_USER_AGENT);
         }
 
         /// <summary>
