+++ src/Shared-lib/PackageManagers/CPANProjectManager.cs
@@ -42,7 +42,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
                 return false;
             }
             string packageName = purl.Name;
-            using HttpClient httpClient = CreateHttpClient();
+            HttpClient httpClient = CreateHttpClient();
             return await CheckHttpCacheForPackage(httpClient, $"{ENV_CPAN_ENDPOINT}/release/{packageName}", useCache);
         }
 
@@ -65,7 +65,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
                 return downloadedPaths;
             }
             // Locate the URL
-            using HttpClient httpClient = CreateHttpClient();
+            HttpClient httpClient = CreateHttpClient();
             string? packageVersionUrl = null;
             string? html = await GetHttpStringCache(httpClient, $"{ENV_CPAN_ENDPOINT}/release/{packageName}");
             HtmlParser parser = new();
@@ -149,7 +149,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
             {
                 string packageName = purl.Name;
                 List<string> versionList = new();
-                using HttpClient httpClient = CreateHttpClient();
+                HttpClient httpClient = CreateHttpClient();
 
                 string? html = await GetHttpStringCache(httpClient, $"{ENV_CPAN_ENDPOINT}/release/{packageName}");
                 HtmlParser parser = new();
@@ -186,7 +186,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
                 string? packageName = purl.Name;
                 if (packageName != null)
                 {
-                    using HttpClient httpClient = CreateHttpClient();
+                    HttpClient httpClient = CreateHttpClient();
                     string? contentRelease = await GetHttpStringCache(httpClient, $"{ENV_CPAN_ENDPOINT}/release/{packageName}");
                     string? contentPod = await GetHttpStringCache(httpClient, $"{ENV_CPAN_ENDPOINT}/pod/{packageName.Replace("-", "::")}");
                     return contentRelease + "\n" + contentPod;
