+++ src/Shared-lib/PackageManagers/MavenProjectManager.cs
@@ -49,7 +49,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
                 foreach (string suffix in suffixes)
                 {
                     string url = $"{ENV_MAVEN_ENDPOINT}/{packageNamespace}/{packageName}/{packageVersion}/{packageName}-{packageVersion}{suffix}.jar";
-                    using HttpClient httpClient = CreateHttpClient();
+                    HttpClient httpClient = CreateHttpClient();
 
                     System.Net.Http.HttpResponseMessage result = await httpClient.GetAsync(url);
                     result.EnsureSuccessStatusCode();
@@ -98,7 +98,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
             }
             string packageNamespace = purl.Namespace.Replace('.', '/');
             string packageName = purl.Name;
-            using HttpClient httpClient = CreateHttpClient();
+            HttpClient httpClient = CreateHttpClient();
 
             return await CheckHttpCacheForPackage(httpClient, $"{ENV_MAVEN_ENDPOINT}/{packageNamespace}/{packageName}/maven-metadata.xml", useCache);
         }
@@ -114,7 +114,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
             {
                 string packageNamespace = purl.Namespace.Replace('.', '/');
                 string packageName = purl.Name;
-                using HttpClient httpClient = CreateHttpClient();
+                HttpClient httpClient = CreateHttpClient();
 
                 string? content = await GetHttpStringCache(httpClient, $"{ENV_MAVEN_ENDPOINT}/{packageNamespace}/{packageName}/maven-metadata.xml");
                 List<string> versionList = new();
@@ -147,7 +147,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
             {
                 string? packageNamespace = purl?.Namespace?.Replace('.', '/');
                 string? packageName = purl?.Name;
-                using HttpClient httpClient = CreateHttpClient();
+                HttpClient httpClient = CreateHttpClient();
                 if (purl?.Version == null)
                 {
                     foreach (string? version in await EnumerateVersions(purl))
