+++ src/Shared-lib/PackageManagers/VSMProjectManager.cs
@@ -46,7 +46,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
 
             try
             {
-                using HttpClient httpClient = CreateHttpClient();
+                HttpClient httpClient = CreateHttpClient();
                 Stream? resultStream = null;
                 string? cacheResult = GetCache(packageName);
                 if (cacheResult != null)
@@ -189,7 +189,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
                     requestMessage.Headers.Add("Accept", "application/json;api-version=3.0-preview.1");
                     string postContent = $"{{filters:[{{criteria:[{{filterType:7,value:\"{packageName}\"}}],pageSize:1000,pageNumber:1,sortBy:0}}],flags:131}}";
                     requestMessage.Content = new StringContent(postContent, Encoding.UTF8, "application/json");
-                    using HttpClient httpClient = CreateHttpClient();
+                    HttpClient httpClient = CreateHttpClient();
 
                     HttpResponseMessage response = await httpClient.SendAsync(requestMessage);
                     resultStream = await response.Content.ReadAsStreamAsync();
@@ -248,7 +248,7 @@ namespace Microsoft.CST.OpenSource.PackageManagers
         {
             try
             {
-                using HttpClient httpClient = CreateHttpClient();
+                HttpClient httpClient = CreateHttpClient();
 
                 return await GetHttpStringCache(httpClient, $"{ENV_VS_MARKETPLACE_ENDPOINT}/items?itemName={purl.Namespace}/{purl.Name}");
             }
