+++ src/oss-tests/DetectCryptographyTests.cs
@@ -31,7 +31,6 @@ namespace Microsoft.CST.OpenSource.Tests
         [DataRow("pkg:npm/sm4-demo", "Cryptography.Implementation.BlockCipher.SM4")]
         public async Task TestPackageDetectionSucceeds(string purl, params string[] expectedTags)
         {
-            Console.WriteLine("passed this line");
             await TestDetectCryptography(purl, expectedTags);
         }
